stream length can be indirect object
