+++ pdfreader/document.py
@@ -1,9 +1,8 @@
 import logging
 
-from .exceptions import ParserException
 from .registry import Registry
 from .parsers import RegistryPDFParser
-from .types import null, IndirectObject, Stream, Array, Dictionary, IndirectReference, obj_factory
+from .types import IndirectObject, Stream, Array, Dictionary, IndirectReference, obj_factory
 
 
 class PDFDocument(object):
@@ -25,12 +24,8 @@ class PDFDocument(object):
 
         self.registry = Registry()
         self.parser = RegistryPDFParser(fobj, self.registry)
-        self.header = self.parser.pdf_header()
-        self.trailer = self.parser.pdf_trailer()
-
-        # save initial state for brute-force objects lookup
-        self.parser.reset(self.header.offset)
-        self.brute_force_state = self.parser.get_state()
+        self.header = self.parser.header
+        self.trailer = self.parser.trailer
 
         #: references to document's Catalog instance
         self.root = self.obj_by_ref(self.trailer.root)
@@ -78,64 +73,8 @@ class PDFDocument(object):
             on_return()
         return obj
 
-    def locate_object(self, num, gen):
-        """
-        Locates an object by it's number and generation.
-
-        Objects lookup order:
-          #. Known objects in registry (located before)
-          #. XRef tables lookups
-          #. Brute-force reading objects one by one from file start
-
-        :param num: object number
-        :type num: int
-
-        :param gen: object generation
-        :type gen: int
-
-        :return: instance of one of supported PDF types (incl. null object) if found, null object otherwise.
-                 Doesn't resolve indirect references.
-        """
-        # locate in registry
-        if self.registry.is_registered(num, gen):
-            return self.registry.get(num, gen)
-
-        # Locate by xref
-        for xref in self.trailer.xrefs:
-            # try to find in-use object
-            xre = xref.in_use.get(num)
-            if xre and xre.generation == gen:
-                try:
-                    self.parser.reset(xre.offset)
-                    self.parser.indirect_object()
-                except ParserException:
-                    pass
-                if self.registry.is_registered(num, gen):
-                    break
-            # Try to find a compressed object
-            xre = xref.compressed.get(num)
-            if xre and xre.generation == gen:
-                # Need to locate Object Stream in order to locate a compressed object
-                if xre.number != num:
-                    # Avoid infinite loops when compressed object is not listed on xref
-                    # See https://github.com/maxpmaxp/pdfreader/issues/16
-                    self.locate_object(xre.number, xre.generation)
-                    if self.registry.is_registered(num, gen):
-                        break
-
-        while not self.registry.is_registered(num, gen):
-            try:
-                _ = self.next_brute_force_object()
-            except ParserException:
-                # treat not-found objects as nulls
-                logging.exception("!!!Failed to locate {} {}: assuming null".format(num, gen))
-                self.registry.register(IndirectObject(num, gen, null))
-                break
-        obj = self.registry.get(num, gen)
-        return obj
-
     def obj_by_ref(self, objref):
-        obj = self.locate_object(objref.num, objref.gen)
+        obj = self.parser.locate_object(objref.num, objref.gen)
         return obj_factory(self, obj)
 
     def deep_obj_by_ref(self, obj, maxdepth=100):
@@ -146,13 +85,6 @@ class PDFDocument(object):
 
         if isinstance(obj, IndirectObject):
             raise ValueError("Max reference depth exceeded")
-        return obj
-
-    def next_brute_force_object(self):
-        self.parser.set_state(self.brute_force_state)
-        self.parser.maybe_spaces_or_comments()
-        obj = self.parser.body_element() # can be either indirect object, startxref or trailer
-        self.brute_force_state = self.parser.get_state() # save state for the next BF
         return obj
 
     def pages(self):
